Apply merge results from the previous repo

+++ src/vehiclelist.h
@@ -32,9 +32,9 @@ struct VehicleListIdentifier {
 	CompanyID company;    ///< The company associated with this list.
 	uint32 index;         ///< A vehicle list type specific index.
 
-	uint32 Pack() const;
-	bool UnpackIfValid(uint32 data);
-	static VehicleListIdentifier UnPack(uint32 data);
+	uint64 Pack() const;
+	bool UnpackIfValid(uint64 data);
+	static VehicleListIdentifier UnPack(uint64 data);
 
 	bool Valid() const { return this->type < VLT_END; }
